+++ libs/iresearch/include/iresearch/store/fs_directory.cpp
@@ -298,7 +298,7 @@ class FsIndexInput : public BufferedIndexInput {
     byte_type buf[sizeof _buf];
 
     for (auto pos = begin; pos < end;) {
-      const auto to_read = std::min(end - pos, sizeof buf);
+      const auto to_read = std::min(end - pos, static_cast<uint64_t>(sizeof buf));
       pos += const_cast<FsIndexInput*>(this)->ReadInternal(buf, to_read);
       crc.process_bytes(buf, to_read);
     }
