+++ libs/iresearch/include/iresearch/store/store_utils.cpp
@@ -64,8 +64,11 @@ void BytesViewInput::ReadBytes(bstring& buf, size_t size) {
 uint32_t BytesViewInput::Checksum(uint64_t offset) const {
   Crc32c crc;
 
-  crc.process_bytes(_pos,
-                    std::min((_pos - _data.data()) + offset, _data.size()));
+  crc.process_bytes(
+    _pos,
+    std::min(
+      (_pos - _data.data()) + offset,
+      static_cast<uint64_t>(_data.size())));
 
   return crc.checksum();
 }
